ByteCraft_Challenge/auth/auth_app/requirements.txt
asgiref==3.7.2
attrs==23.1.0
black==23.9.1
certifi==2023.7.22
charset-normalizer==3.2.0
click==8.1.7
colorama==0.4.6
coverage==7.3.1
discord-rich-presence==1.1.0
Django==4.2.5
django-cors-headers==4.2.0
django-extensions==3.2.1
django-filter==22.1
django-js-asset==2.1.0
django-mptt==0.14.0
django-pytest==0.2.0
djangorestframework==3.14.0
djangorestframework-jsonapi==6.0.0
djangorestframework-simplejwt==5.3.0
drf-spectacular==0.26.4
flake8==6.1.0
idna==3.4
inflection==0.5.1
iniconfig==2.0.0
jsonschema==4.19.1
jsonschema-specifications==2023.7.1
mccabe==0.7.0
mypy-extensions==1.0.0
packaging==23.1
pathspec==0.11.2
platformdirs==3.10.0
pluggy==1.3.0
psycopg2==2.9.9
psycopg2-binary==2.9.7
pycodestyle==2.11.0
pyflakes==3.1.0
PyJWT==2.8.0
pytest==7.4.2
pytest-cov==4.1.0
pytest-django==4.5.2
python-dotenv==0.21.0
pytz==2022.6
PyYAML==6.0.1
redis==5.0.0
referencing==0.30.2
requests==2.31.0
rpds-py==0.10.3
six==1.16.0
spotipy==2.23.0
sqlparse==0.4.3
tzdata==2022.6
uritemplate==4.1.1
urllib3==2.0.5
